requirements.txt
scapy
python-nmap
matplotlib
networkx
requests
